fix: rename category entity to aggregate
  using ddd concepts
