+++ src/@shared/src/domain/models/entities/aggregate-root.ts
@@ -1,0 +1,4 @@
+import { EntityBase } from './entity-base';
+
+// agregado lida com algo que entidades não lidam, eventos de domínio
+export abstract class AggreateRoot extends EntityBase {}
