+++ src/core/data/use-cases/category/update-category/update-category.use-case.ts
@@ -5,7 +5,10 @@ import {
   IUpdateCategoryInput,
   IUpdateCategoryUseCase,
 } from 'src/core/domain/contracts/use-cases/category/update/update-category';
-import { Category, CategoryId } from 'src/core/domain/entities/category.entity';
+import {
+  Category,
+  CategoryId,
+} from 'src/core/domain/entities/category.aggregate';
 import {
   CategoryOutput,
   CategoryOutputMapper,
